ROBOT_DEMO: change lift to crane

+++ orte/Robot_Demo/src/org/ocera/orte/demo/MagnetCmdPublish.java
@@ -65,7 +65,7 @@ public class MagnetCmdPublish {
 		}
 	}
 	
-	public void send(short magnet) {
+	public void send(int magnet) {
 		magnetmsg.magnet = magnet;
 		pub.send(magnetmsg);
 	}
